requirements.txt
aiohttp
loguru
python-telegram-bot